Move split_grid and merge_grids to separate file in tools
